Move OpenGraph image proxying to app layer
